move sample set functions to sample.py
