+++ sim/sample.py
@@ -2,6 +2,7 @@ from utils import run
 import errno
 import math 
 import os
+import glob
 
 class Sample:
 
@@ -176,3 +177,42 @@ class RandomSample(Sample):
                     "-indel_count", str(self.num_indels),
                     "-seed", str(self.seed)]
         self._simulate_genome_base(reference_path, simulated_genome_path, params)
+
+
+def vcf_samples(datasets, snippy_dir='/mnt/fsx-027/snippy/'):
+    vcf_filepaths = []
+    for set_i in datasets:
+        vcf_dir = snippy_dir + set_i
+        if not os.path.isdir(vcf_dir):
+            raise Exception("Predefined SNP directory not found") 
+        vcf_dir = os.path.join(vcf_dir, '')
+        vcf_filepaths.extend(glob.glob(vcf_dir+'*.vcf'))
+    
+    # sort vcf_filepaths for using consistent seed values accross runs
+    vcf_filepaths.sort()
+    samples = []
+    seed_value = 0
+    for filepath in vcf_filepaths:
+        seed_value+=1 # different seed value for each sample
+        samples.append(VcfSample(filepath, seed=seed_value, per_base_error_rate="0.001-0.01"))    
+
+    return samples
+
+def quick_samples():
+    return [RandomSample(seed=1)]
+
+def random_samples():
+    return [RandomSample(seed=1, per_base_error_rate="0.001-0.01"),
+            RandomSample(seed=666, per_base_error_rate="0.001-0.01")]
+
+def standard_samples():
+    return vcf_samples(('standard',)) + random_samples()
+
+def aph_samples():
+    return vcf_samples(('aph',))
+
+def zwyer_samples():
+    return vcf_samples(('zwyer',))
+
+def all_samples():
+    return vcf_samples(('aph', 'zwyer')) + random_samples()
